Improvements & for text validation use en.json
