+++ cypress/integration/2.1.3-verify_review_with_signed_user.spec.js
@@ -1,12 +1,11 @@
+import { loginViaCookies, preserveCookie } from '../support/common/support.js'
+import { testCase1 } from '../support/outputvalidation.js'
 import {
-  loginViaCookies,
-  preserveCookie,
-  updateRetry,
-} from '../support/common/support.js'
-import { testCase1 } from '../support/review_and_ratings.outputvalidation.js'
-import { verifyUserIsNotAbletoAddReviewAgain } from '../support/testcase.js'
-import rrselectors from '../support/reviews_and_ratings.selectors.js'
-import { updateSettings } from '../support/review_and_ratings_settings.js'
+  verifyUserIsNotAbletoAddReviewAgain,
+  verifyAverageRatings,
+  addedReviewShouldShowToTheUser,
+} from '../support/testcase.js'
+import { updateSettings } from '../support/graphql_testcase.js'
 
 const { product, user1, title, configuration } = testCase1
 
@@ -17,20 +16,9 @@ describe(title, () => {
   describe(`Verify review with Signed In User`, () => {
     loginViaCookies()
 
-    it(
-      'Added review should show immediately to the user',
-      updateRetry(2),
-      () => {
-        cy.openStoreFront(true)
-        cy.openProduct(product, true)
-        cy.get(rrselectors.PostalCode, { timeout: 20000 }).should('be.visible')
-        cy.get('span[class*=reviewComment]').contains(user1.line)
-      }
-    )
+    addedReviewShouldShowToTheUser(product, user1.line)
 
-    it('Verify average ratings', () => {
-      cy.getAverageRating(user1, product)
-    })
+    verifyAverageRatings(product, user1)
 
     verifyUserIsNotAbletoAddReviewAgain()
