+++ cypress/integration/2.2.2-anonymous_user.spec.js
@@ -1,7 +1,7 @@
 import { loginViaCookies } from '../support/common/support.js'
 import { restrictAnonymousUser } from '../support/testcase.js'
-import { updateSettings } from '../support/review_and_ratings_settings.js'
-import { testCase2 } from '../support/review_and_ratings.outputvalidation'
+import { updateSettings } from '../support/graphql_testcase.js'
+import { testCase2 } from '../support/outputvalidation.js'
 
 const { title, configuration, product } = testCase2
