+++ cypress/integration/2.4.1-anonymous_user.spec.js
@@ -1,13 +1,13 @@
 import { loginViaCookies } from '../support/common/support.js'
-import { testCase4 } from '../support/review_and_ratings.outputvalidation'
-import { updateSettings } from '../support/review_and_ratings_settings.js'
+import { testCase4 } from '../support/outputvalidation.js'
+import { updateSettings } from '../support/graphql_testcase.js'
 import {
   restrictAnonymousUser,
   verifiedReviewTestCase,
   reviewTestCase,
   verifySorting,
 } from '../support/testcase.js'
-import { reviewsAndRatingsConstants } from '../support/reviews_and_ratings.constants.js'
+import { reviewsAndRatingsConstants } from '../support/constants.js'
 
 const { title, configuration, product } = testCase4
