+++ cypress/integration/2.7.1-disable_display_settings_&_verify_with_anonymous_user.spec.js
@@ -1,6 +1,6 @@
 import { loginViaCookies } from '../support/common/support.js'
-import { updateSettings } from '../support/review_and_ratings_settings.js'
-import { verifySettings } from '../support/review_and_ratings.js'
+import { updateSettings } from '../support/graphql_testcase.js'
+import { verifySettings } from '../support/testcase.js'
 
 describe('Testing as anonymous user -> Disable display setting', () => {
   loginViaCookies({ storeFrontCookie: false })
