+++ cypress/integration/2.7.4-disable_display_settings_&_verify_with_signed_in_user.spec.js
@@ -1,6 +1,6 @@
 import { preserveCookie, loginViaCookies } from '../support/common/support.js'
-import { updateSettings } from '../support/review_and_ratings_settings.js'
-import { verifySettings } from '../support/review_and_ratings.js'
+import { updateSettings } from '../support/graphql_testcase.js'
+import { verifySettings } from '../support/testcase.js'
 
 describe('Disable display setting and test', () => {
   loginViaCookies()
